Refaktorert joark-service og -klient
