--- src/main/kotlin/no/nav/tiltakspenger/soknad/api/joark/Joark.kt
@@ -1,5 +1,0 @@
-package no.nav.tiltakspenger.soknad.api.joark
-
-interface Joark {
-    suspend fun opprettJournalpost(dokumentInnhold: Journalpost, callId: String): String // JournalpostId
-}
